+++ packages/expo-video/android/src/main/java/expo/modules/video/player/VideoPlayer.kt
@@ -20,6 +20,7 @@ import androidx.media3.exoplayer.DecoderReuseEvaluation
 import androidx.media3.exoplayer.DefaultRenderersFactory
 import androidx.media3.exoplayer.ExoPlayer
 import androidx.media3.exoplayer.analytics.AnalyticsListener
+import androidx.media3.exoplayer.hls.HlsManifest
 import androidx.media3.exoplayer.trackselection.DefaultTrackSelector
 import androidx.media3.session.MediaSession
 import androidx.media3.ui.PlayerView
@@ -237,7 +238,9 @@ class VideoPlayer(val context: Context, appContext: AppContext, source: VideoSou
       val newAudioTracks = audioTracks.availableAudioTracks
       val newCurrentSubtitleTrack = subtitles.currentSubtitleTrack
       val newCurrentAudioTrack = audioTracks.currentAudioTrack
-      availableVideoTracks = tracks.toVideoTracks()
+      val hlsManifest = player.currentManifest as? HlsManifest
+
+      availableVideoTracks = tracks.toVideoTracks(hlsManifest)
 
       if (isLoadingNewSource) {
         sendEvent(
@@ -539,17 +542,23 @@ class VideoPlayer(val context: Context, appContext: AppContext, source: VideoSou
 // Extension functions
 
 @OptIn(UnstableApi::class)
-private fun Tracks.toVideoTracks(): List<VideoTrack> {
+private fun Tracks.toVideoTracks(sourceManifest: HlsManifest?): List<VideoTrack> {
   val videoTracks = mutableListOf<VideoTrack?>()
   for (group in this.groups) {
     for (i in 0 until group.length) {
       val format = group.getTrackFormat(i)
       val isSupported = group.isTrackSupported(i)
+      val hlsVariant = sourceManifest?.multivariantPlaylist?.variants?.firstOrNull {
+        it.format.id == format.id
+      }
+
+      // We provide the variant url only for HLS sources
+      val variantUrl = hlsVariant?.url
 
       if (!MimeTypes.isVideo(format.sampleMimeType)) {
         continue
       }
-      videoTracks.add(VideoTrack.fromFormat(format, isSupported))
+      videoTracks.add(VideoTrack.fromFormat(format, isSupported, variantUrl))
     }
   }
   return videoTracks.filterNotNull()